Add word warp file processing tool
This tool checks a file of 6-letter "word warp" words for correctness.
While playing word warp, I maintain a file of all the 6-letter words
(and their anagrams) which I've seen.  The file has a particluar
structure, and this program verifies the structure.  Eventually, it
could also show some statistics about the words in the file.